Add files via upload
so the screenshot is from the project and is what needs to be added to the report. everything underlined is what still needs to be added and so if wanted to add to it, feel free. this also is not a final draft so if you want to reword anything, feel free to do that as well!